a more reusable implementation of the memcached aio backend
