+++ django/core/cache/backends/base.py
@@ -280,3 +280,35 @@ class BaseCache:
     def close(self, **kwargs):
         """Close the cache connection"""
         pass
+
+
+class BaseAioCache(BaseCache):
+    "An implementation of a cache binding using aiocache"
+    def __init__(self, server, params):
+        super().__init__(server, params, library=self.get_library_class(),
+                         value_not_found_exception=ValueError)
+
+    def get_library_class(self):
+        raise NotImplementedError('subclasses of BaseAioCache must implement a get_library_class method that returns a backend supported by aiocache')
+
+    def get_serializer_class(self):
+        from aiocache.serializers import PickleSerializer
+        return PickleSerializer
+
+    @property
+    def _cache(self):
+        if getattr(self, '_client', None) is None:
+            host, port = self._servers[0].split(':')
+            self._client = self._lib(endpoint=host, port=port,
+                                     serializer=self.get_serializer_class()())
+        return self._client
+
+    async def get(self, key, default=None, version=None):
+        key = self.make_key(key, version=version)
+        return await self._cache.get(key, default=default)
+
+    async def set(self, key, value, timeout=DEFAULT_TIMEOUT, version=None):
+        key = self.make_key(key, version=version)
+        if not await self._cache.set(key, value, self.get_backend_timeout(timeout)):
+            # make sure the key doesn't keep its old value in case of failure to set (memcached's 1MB limit)
+            await self._cache.delete(key)
