+++ django/core/cache/backends/memcached.py
@@ -4,7 +4,7 @@ import pickle
 import re
 import time
 
-from django.core.cache.backends.base import DEFAULT_TIMEOUT, BaseCache
+from django.core.cache.backends.base import DEFAULT_TIMEOUT, BaseCache, BaseAioCache
 from django.utils.asyncio import auto_async, async_unsafe
 from django.utils.functional import cached_property
 
@@ -207,28 +207,8 @@ class PyLibMCCache(BaseMemcachedCache):
         pass
 
 
-class AioMemcacheCache(BaseMemcachedCache):
-    "An implementation of a cache binding using aiocache"
-    def __init__(self, server, params):
+class AioMemcachedCache(BaseAioCache, BaseMemcachedCache):
+    "An implementation of an async (mem)cache binding using aiocache"
+    def get_library_class(self):
         import aiocache
-        super().__init__(server, params, library=aiocache.MemcachedCache,
-                         value_not_found_exception=ValueError)
-
-    @property
-    def _cache(self):
-        if getattr(self, '_client', None) is None:
-            from aiocache.serializers import PickleSerializer
-            host, port = self._servers[0].split(':')
-            self._client = self._lib(
-                endpoint=host, port=port, serializer=PickleSerializer())
-        return self._client
-
-    async def get(self, key, default=None, version=None):
-        key = self.make_key(key, version=version)
-        return await self._cache.get(key, default=default)
-
-    async def set(self, key, value, timeout=DEFAULT_TIMEOUT, version=None):
-        key = self.make_key(key, version=version)
-        if not await self._cache.set(key, value, self.get_backend_timeout(timeout)):
-            # make sure the key doesn't keep its old value in case of failure to set (memcached's 1MB limit)
-            await self._cache.delete(key)
+        return aiocache.MemcachedCache
